requirements.txt
wheel
blinker
chardet
click
dash
flask-assets
flask-dance
flask-login
Flask-SQLAlchemy
flask
google-auth
google-auth-oauthlib
idna
itsdangerous
Jinja2
MarkupSafe
numpy
oauthlib
pandas
plotly
pyasn1
pyasn1-modules
python-dotenv
requests
urllib3
uwsgi
##Before OAuth
easydev
colormap
gevent
dateparser
pyarrow
gunicorn>=19.7.1
dash-bootstrap-components==0.13.1
dask[complete]
fastparquet
dash-renderer>=1.6.0
dash-html-components>=1.0.3
dash-core-components>=1.10.2
numpy
pandas
flask_caching
joblib
matplotlib
statsmodels
nbconvert
ipykernel
jupyter_client
sphinx-gallery
zipp
geopandas
shapely
dash_daq
